# Conflicts:
#	bundles/CoreBundle/Resources/translations/en.json

+++ models/DataObject/Fieldcollection/Dao.php
@@ -212,7 +212,7 @@ class Dao extends Model\Dao\AbstractDao
             }
         }
 
-        if (!$this->model->isFieldDirty('_self')) {
+        if (!$this->model->isFieldDirty('_self') && !DataObject\AbstractObject::isDirtyDetectionDisabled()) {
             return false;
         }
         $whereLocalizedFields = "(ownertype = 'localizedfield' AND "
@@ -221,7 +221,7 @@ class Dao extends Model\Dao\AbstractDao
             . ' AND ' . $this->db->quoteInto('src_id = ?', $object->getId()). ')';
 
         if ($saveMode) {
-            if (!$this->model->hasDirtyFields() && $hasLocalizedFields) {
+            if (!DataObject\AbstractObject::isDirtyDetectionDisabled() && !$this->model->hasDirtyFields() && $hasLocalizedFields) {
                 // always empty localized fields
                 $this->db->deleteWhere('object_relations_' . $object->getClassId(), $whereLocalizedFields);
